requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
pyyaml>=5.4.0
tqdm>=4.62.0
pytest>=6.2.0